Fix (ref: T29168): remove pipeline related parameters and methods from globalConfig
- pipeline related parameters and methods are stored seperatly in AiPipelineConfiguration.php
